libcamera_timezone: fix a crash caused by destruction of TimezoneMap
The AtExitManager would try to run the destruction after dlclose(),
which cause a segfault.

Our C++ style guide discourage this kind of usage but the linter will
not catch it. The new base/no_destructor.h is the preferred solution but
still not available in CrOS, so we use a simpler solution here with the
cost of a pointer indirection and a malloc.

refs:
https://chromium.googlesource.com/chromium/src/base/+/master/no_destructor.h
go/totw/110 (internal only)

BUG=b:74577773,b:74098691
TEST=emerge-nautilus cros-camera-libcamera_timezone
TEST=emerge-nautilus cros-camera-hal-usb
TEST=cros_camera_test --camera_hal_path=/usr/lib64/camera_hal/usb.so
    '--gtest_filter=-Camera3DeviceTest/Camera3AlgoSandboxIPCErrorTest*:\
    Camera3RecordingFixture/Camera3BasicRecordingTest*'

Change-Id: I3837bf5b70527b6affb91e85798f198dda265217
Reviewed-on: https://chromium-review.googlesource.com/960183
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Shik Chen <shik@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>