arc: Add exception file for FIFO blocking on stateful.
Symlink and FIFO access on the stateful partition are being blocked by
CL:966683 for security reasons. New file specifies a policy exception
so that FIFOs can be used in the /data directory (under /home).

See b/74868050 for details on the need for the exception.

BUG=chromium:655606
TEST=emerged to board and observed file
Change-Id: I86241ce440254ff81091f3145258a41d561cc130
Reviewed-on: https://chromium-review.googlesource.com/961305
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>