chromeos-dbus-bindings: refactor D-Bus signature parsing
This introduces an intermediate representation for D-Bus type
signatures. This is useful because it allows us to do some extra
inspection of types without having to parse strings that we've
just generated. It also simplifies the parsing code since we can
create the intermediate representation during parsing and leave
the creation of the actual string type until later.

This makes it easier to change the way types are represented as
strings. This will be helpful when we implement generating
different sets of bindings for FDs, and easy separation between
type signatures used to extract arguments from messages and push
arguments into them will be useful for that as well.

BUG=chromium:821633
TEST=unit tests

Change-Id: I42a7b8c345e9df98b8e74b9889bcd1e0d439ea32
Reviewed-on: https://chromium-review.googlesource.com/961720
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>