cryptohome: Refactor the functions to use multiple PCR indexes at input
For additional security in ARC sessions, a new cryptohome key is
required that is bound to both, PCR0 and PCR4. This CL updates the methods
that will be used for creating of the new key, enabling them to accept
multiple PCR indexes and values at the input.

The new key is planned to be created and used in a follow-up set of CLs
that will perform the migration of vault keyset by encrypting it with the
new key.

Desgin doc for proposal:
https://docs.google.com/document/d/1MGmglGvneHMHuL7jIRiTHspSGNXdIhwMMxvnpXwsmVk/edit?ts=5a83f46b#

BUG=chromium:812165
TEST=Unit tests and cryptohome-tpm-live-test on both tpm1.2 and tpm2.0 devices.
CQ-DEPEND=CL:1014038

Change-Id: I76f6fcbe24c490532314b327a5f166f61101225f
Reviewed-on: https://chromium-review.googlesource.com/964362
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>