chromeos-dbus-bindings: clean up MethodNameGenerator
This class has an extra, unused public method whose sole purpose
is to concatenate some strings together, all but one of which are
constants. We then proceed to use this string in a StringPrintf
invocation so we're doing a lot of extra work for nothing. Clean
this up.

BUG=None
TEST=unit tests

Change-Id: I8e30d97f9208c97377257a176ffa0f7cd48254f9
Reviewed-on: https://chromium-review.googlesource.com/967413
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>