chromeos-config: Add FW_NAME build target filter.
For development purposes, it is convenient to only build a subset of
build targets for a given config. This change allows users to set the
FW_NAME environment variable to only build for one named set of build
targets at a time. The name is currently keyed off of the coreboot image
name, but should be migrated to a "name" property later.

BUG=b:74436651
TEST=unittests, manual runs on custom dtsi file

Change-Id: Ie3c5aad28fdf184d6e531cfb950c14a6da97b95f
Reviewed-on: https://chromium-review.googlesource.com/971741
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>