debugd: Add netstat and logcat to log_tool (chrome://system)
Add netstat so that users can see open sockets.  Since we are adding
more and more services that listen on open ports (arc-networkd,
vm_concierge, etc.) this can be useful to troubleshoot why a Chrome
app is unable to bind to a port.  log_tool already shows ps and
ifconfig, so there is precedent for offering this level of information.

Add logcat to aid in debugging Android issues when the user does not
have shell access.

BUG=none
TEST=manually visit chrome://system
TEST=`dbus-send --reply-timeout=30000 --system --print-reply --fixed \
      --dest=org.chromium.debugd /org/chromium/debugd \
      org.chromium.debugd.GetFeedbackLogs`

Change-Id: I2fae76bd43c4bad0f333d2c8bfebdf5556bd0e23
Reviewed-on: https://chromium-review.googlesource.com/972356
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>