cryptohome: refactor mount-encrypted error reporting
Switch from booleans/ints for returning status from internal functions
to an enum of result codes to allow future extension of the set of
distinct error types. This unifies the style of reporting errors
across mount-encrypted: before the change some functions were using
0 for success, amnd some weer using 1. The new result codes are also
returned from mount-encrypted executable as exit codes to allow the
callers implement different reactions to different types of errors.

This CL doesn't change any functionality or add new types of errors.
In one case, an impossible branch  - a check if the result was success
after we have already checked that it was an error - is eliminated to
slightly simplify the code.

BUG=chromium:702724
TEST=build, boot

Change-Id: If465fa65c99c61f99e3b1b0bbf9d90af6b1507fa
Reviewed-on: https://chromium-review.googlesource.com/972340
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>