arc: Modify exception file for FIFO blocking on stateful.
Symlink and FIFO access on the stateful partition are being blocked by
CL:966683 for security reasons. We already attempted to add an exception
file for this policy for the /home/root subtree, so that the Android
/data dir could use FIFOs. Turns out /home/root/$hash is a bind-mount of
/home/.shadow/$hash/mount/root, meaning the FIFOs are used under
/home/.shadow, not /home/root. Backing this exception one step up the
tree to /home allows Android to use FIFOs in the /data directory.

See b/76173776 for details.

BUG=chromium:655606
TEST=ran cheets_CTS_N.7.1_r15.arm.CtsLibcoreTestCases (works for x86 as
well) and obseved FIFO blocking in existing code, but no FIFO blocking
with this change.

Change-Id: Ia91ca8fc350730b65ec682688ad455542236bb2f
Reviewed-on: https://chromium-review.googlesource.com/978661
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>