vm_tools: Add LaunchContainerApplication call to concierge
This adds the D-Bus call to concierge for launching an application in a
container. This also removes the token->ip mapping that we had and
instead replaces the IP with a gRPC stub that we can use for executing
RPCs into the container.

BUG=None
TEST=Verified works with concierge_client
CQ-DEPEND=CL:980665, CL:980667

Change-Id: I4ee229f74b9164a9a505c4048ea7275b64fd3f11
Reviewed-on: https://chromium-review.googlesource.com/980666
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>