login: Add EmitAshInitialized dbus method
Adds EmitAshInitialized method to session manager interface.
The method emits ash-initialized upstart signal on behalf of
the browser. It is expected to be called when the ash shell is
initialized.

The signal will be used to run jobs that depend on ash shell - for
example to run hammerd. The hammerd job should run when ash is
initialized to ensure ash receives dbus signals emitted by hammerd
that match the current hammer state - this will ensure that ash
displays notifications matching the current hammer state, if needed.

BUG=chromium:818057
TEST=(once Chrome and hammerd changes have landed) While on login
     screen, attach a detachable base different than the last one
     used by the user focused on the login screen. Run "restart ui".
     Upon Chrome restart, a warning that keyboard is different than
     the last one used by the current user should be displayed.

Change-Id: Ib42def82310c72e37c2476f39c8e121f8d4d09f4
Reviewed-on: https://chromium-review.googlesource.com/981604
Commit-Ready: Toni Barzic <tbarzic@chromium.org>
Tested-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>