vm_tools: Add sending application list from garcon
This sends a list of all the existing .desktop files to concierge over
gRPC.  It does some filtering on the list before sending it. It also
sets up a filesystem watcher so that it can be notified of any changes
to the files and then send an update to concierge with the new
information.

BUG=None
TEST=Unit tests pass, verified with manual testing
CQ-DEPEND=CL:993401

Change-Id: I66b81480efe06bdf7456bd8d308e4e834f9cc72f
Reviewed-on: https://chromium-review.googlesource.com/993402
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>