power: Use simple exponential smoothing for light level.
Use the following formula to do light level smoothing.
nex_lux = alpha * lux_now + (1 - alpha) * lux_previous

The alpha (aka smoothing constant) is default to be 1.0
which means smoothing is disabled. This can be override
for each platform using newly created powerd pref called
|als_smoothing_constant|.

With reasonable smoothing constant like 0.2, this will
makes screen to not adjusting brightness for sequence of
ALS readings like [120, 115, 5, 3, 5, 120, 115] which is easily
occured when something covers the ALS sensor for a short time.

BUG=chromium:826968
TEST=newly created unit test, manual test.
eve won't dim screen if cover ALS for a short time with alpha=0.2

Change-Id: I32153eabdceaf685c38e28f02020e13fcde7c407
Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/985118