cryptohome: tpm2: revert GetEndorsementPublicKey to not implemented
CL:827064 incorrectly implemented Tpm2Impl::GetEndorsementPublicKey().
It should return a DER encoded EK, not EK modulus. Now with CL:968869
renaming the underlying trunks function this incorrect implementation
breaks the build. Nobody is using this function for tpm2, so it's
sufficient to revert it back to LOG(ERROR) << "Not implemented" as it
was pre CL:827064.

BUG=chromium:822426
BUG=chromium:826842
TEST=unit tests

Change-Id: I0594a30822eaa5ef74a62f5c8b189fa0fea77f66
Reviewed-on: https://chromium-review.googlesource.com/985493
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>