chromeos-config: Prep for native YAML impl
YAML was highly coupled to the DTS details and nothing could really be
leveraged from a YAML perspective.

This is the first in a series of steps to cleanly decouple the YAML from
the DTS implementation at the interface layer.

This will allow progress on v2 without the v1 issues.

BUG=b:66950157
TEST=unit

Change-Id: I4490c26649f5c95cadae14e14d296659693a63b5
Reviewed-on: https://chromium-review.googlesource.com/986641
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Gregory Meinke <gmeinke@chromium.org>