cryptohome: Add GYP deps for code-generated targets
Now the targets that create code-generated files provide the "deps"
specified for all dependencies of the output files. These "deps"
are set up to be propagated to the dependent targets via
"all_dependent_settings".

In particular, this change is for these targets:
* Protobuf bindings generation.
  For these, 'protobuf' is in their own deps because the symbols
  from it are used in the compiled object files.
* D-Bus bindings and adaptors generation.
  As these only generate headers and don't produce any object
  files, the libraries included from these headers are specified
  only as deps for the dependent settings, but not for the targets
  themselves.

Note that the removal of unnecessary "deps" for dependent targets
will be done in a follow-up CL.

BUG=chromium:827221
TEST=none

Change-Id: Iffeab526a66b42ec62f6b7a23b6f59657164869f
Reviewed-on: https://chromium-review.googlesource.com/986436
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>