cryptohome: add --input and --output options
The --input and --output options allows one to separate the input and
output files. This is useful for instance when debugging so as not to
clobber CA responses.

The cryptohome command will read input files using either the value of
--input or that of --file if --input is not provided. It will write
output files using either the value of --output or that of --file if
--output is not provided. This affords backwards compatibility with
current usage.

BUG=chromium:822758
TEST=manual

Change-Id: I8cd98f25ab19eaf2f535527ec4d1d8d65cc4b052
Reviewed-on: https://chromium-review.googlesource.com/987033
Commit-Ready: Yves Arrouye <drcrash@chromium.org>
Tested-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>