vm_tools: concierge: Persist used vsock cids
There is some state in the vhost_vsock kernel subsystem that is not
being cleaned up properly when the crosvm process exits.  This means
that if a user logs out and logs back in, their first few attempts to
start a VM will fail because we reuse vsock context ids.

Save the next context id in a file in /run so that we will never reuse
context ids when vm_concierge restarts.

This change should be reverted once we fix the underlying problem in the
kernel.

BUG=chromium:821478
TEST='vmc start gentoo', 'stop vm_concierge', 'vmc start gentoo',
     log out, log back in, 'vmc start gentoo' works.  Also manually
     inspecting the file with 'xxd /run/vm/next_cid' to ensure that the
     values are being properly persisted.

Change-Id: Id4d13cfe5247cb79a44f63df69127c523db3d3cd
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/987389
Reviewed-by: Stephen Barber <smbarber@chromium.org>