chromeos-config: Fix firmware YAML format
Because we want to share the image in the firmware shellball,
it makes sense to break up the image details itself from the
signing config.  This way, we don't have to override in the
yaml and it's really easy to deal with the shared fw instances.

BUG=b:66950157
TEST=unit

Change-Id: I1eb3e7c93bf8a7e7e628235494aae26a3324058d
Reviewed-on: https://chromium-review.googlesource.com/988415
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Gregory Meinke <gmeinke@chromium.org>