virtual_file_provider: elide dbus::FileDescriptor
Appending file descriptors to messages can now be done with bare
ints instead of dbus::FileDescriptors, so just pass the FD from
the ScopedFD instead of creating a whole dbus::FileDescriptor
around it.

It's fine to use a "weak" FD from a base::ScopedFD in this case
because the messages are sent in the function that the ScopedFD
is scoped for.

BUG=b:37434548
TEST=unit tests

Change-Id: I53c0d62e006601c64f799476e21ea0f1e31ae13f
Reviewed-on: https://chromium-review.googlesource.com/988851
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>