debugd: grab all feedback logs when stashing "debug logs"
Various parties like to use the `generate_logs` script as a
one-stop-shop for retrieving various system logs. Unfortunately, the
current logs it captures are woefully inadequate for a variety of cases;
it only captures /var/log/ and some ARC logs (logcat?), but it misses
audio diagnostics, console-ramoops, firmware logs, and plenty more.

Most of this useful type of information is already well-maintained in
the endpoint used for generating feedback reports and similar (LogTool).
Let's make use of that instead, by customizing the existing
"GetAllLogs()" to also get ARC++ logs (which debug_logs_tool.cc already
had special-case code for).

BUG=chromium:701271
TEST=`generate_logs`

Change-Id: I157c27c31884b37d102dea328352829161d293b9
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/991610
Reviewed-by: Ben Chan <benchan@chromium.org>