biod: fix downloading twice an updated template
When a template was updated by a match, the template was downloaded and
saved twice due to the improper use of the 'dirty_list' vector.
Fix the resizing of the vector as we want to reserve capacity rather
than inserting null elements.

BUG=b:74381689
TEST=on Meowth, run a fingerprint match from the lock-screen UI
and no longer see "Retrieve updated template 0" twice in biod log.

Change-Id: I0391d89216b2f0bb5a374a878e02d3cfb3ae44cb
Reviewed-on: https://chromium-review.googlesource.com/992132
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>