shill: Add spawn-in-minijail func to ExternalTask.
This change is in preparation to run the shill process tree as non-root.

l2tpipsec_vpn uses the ExternalTask class to spawn programs and
communicate with them over an RPC interface. Add capability to
ExternalTask to spawn programs in a minijail. Requires passing RPC
connection identifiers on the command line instead of in environment
variables since minijail doesn't support the setting of custom
environment variables for a spawned program.

BUG=chromium:649417
TEST=compiles/works as part of larger shill sandboxing debug CL.

Change-Id: Ic27d04568df857151645473fbe2977e7c96bd2c3
Reviewed-on: https://chromium-review.googlesource.com/992733
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Micah Morton <mortonm@chromium.org>