debugd: generate_logs: complain about excess args
The FlagHelper class is weird, in that it doesn't accept flags with a
space between the flag and arg -- e.g., '--output foo' ignores 'foo'
without complaining, and leaves 'foo' as a leftover arg for the caller
to handle. (The user was supposed to know to type '--output=foo'.)

Instead of improve (and possibly break some users of) FlagHelper, let's
just complain if there are any args leftover. This way `generate_logs
--output /tmp/foo` gives a nice error message instead of generating
/tmp/debug-logs_TIMESTAMP.tgz.

BUG=chromium:701271
TEST=`generate_logs --output /tmp/foo` and
     `generate_logs --output=/tmp/foo`

Change-Id: I8d499052af18bc3b1ca64f975f58e024a86b6627
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/996589
Reviewed-by: Ben Chan <benchan@chromium.org>