cryptohome:  Reset LE credentials on successful Mount
When MountEx completes successfully, all the Low Entropy (LE)
credentials associated with the userid in question should be reset.

This patch adds relevant functions to the Homedirs and Crypto classes to
go through all the LE Credentials and invoke the LECredentialManager's
ResetCredential() routine to unlock a locked out credential.

Additionally,
1) Homedirs::GetValidKeyset() and Mount::DecryptVaultKeyset() both iterate
over all keys of a userid while attempting a decrypt. So, we add logic to
skip LE credentials from these iterations, if no key_label is provided. This
prevents the LE Credentials getting locked out by too many attempts.
2) The code that sets low_entropy_credential flag in the key policy was
moved from Crypto, where it was only run when the keyset was decrypted
and decrypted successfully at that to the points where the keyset is
loaded. This allows GetKeyDataEx to get proper policy for the keyset it
reads.

BUG=chromium:809723
TEST=1. cros_run_unit_tests pass.
     2. cryptohome --action=add_key_ex --key_policy=le
        --user=<user> --new_key_label=<label> succeeds.
     3. cryptohome --action=mount_ex --user=<user>
        --key_label=<label> with wrong PIN
        fails.
     4. Repeat 3. till credential is locked (i.e correct PIN also fails)
     5. cryptohome --action=mount_ex --user=<user>
        with non LE password.
        succeeds
     6. cryptohome --action=unmount
     7. cryptohome --action=mount_ex --user=<user>
        --key_label=<label> with correct PIN.
        suceeeds.
        NOTE: The CLI tests are run with a stub implementation.
Change-Id: If1f18a4b8e2e6075c78d14836376bc5b5e97fb3f
Reviewed-on: https://chromium-review.googlesource.com/997576
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>