vm_tools: StartContainer async startup and return enum
This changes the StartContainer call to always use asynchronous startup
and changes the return protobuf to no longer use the success field but
instead use an enum status value. The D-Bus signal for ContainerStarted
will now only be sent if this indicates the state is STARTING (and this
will soon be extended to also have a signal for when failure occurs).

BUG=None
TEST=Verified proper states with concierge_client
CQ-DEPEND=CL:999961

Change-Id: I05a5af05fa202daf31a6ed93a587ad3f1821c08e
Reviewed-on: https://chromium-review.googlesource.com/1000020
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>