chromeos-config: Make signature_id explicit
I've seen the whitelabel signature convention in 4 different places in
different components across the build/runtime systems.

Changing this to be explicit and set as part of the config.
This will also help in the C code generation (since that is yet another
place that won't need to apply the special convention).

BUG=b:66950157
TEST=unit

Change-Id: I9b61d882505f8bfc5cdcc65730e32b4e5edebcf5
Reviewed-on: https://chromium-review.googlesource.com/1000118
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Lann Martin <lannm@chromium.org>