shill: Don't set RTA_OIF on blackhole routes
RTN_BLACKHOLE routes do not have an output interface, because the
packets are discarded.

On IPv6, the RTA_OIF setting is ignored:

    # ip -6 route add blackhole default table 20 metric 1 dev eth0
    # ip -6 route show table 20
    blackhole default dev lo metric 1  error -22

On IPv4, due to subtle differences in the kernel's routing implementation,
the RTA_OIF attribute generates an error:

    # ip -4 route add blackhole default table 20 metric 1 dev eth0
    RTNETLINK answers: Invalid argument

So this means that RoutingTable::CreateBlackholeRoute() never worked
for IPv4.

RTN_BLACKHOLE routes are not tracked in |RoutingTable::tables_| so
the interface_index can be safely removed without encountering trouble in
RoutingTable::RouteMsgHandler().

BUG=b:72861260
TEST=unit tests

Change-Id: Ibf74c611e103c36be01d289b267836835d64ac74
Reviewed-on: https://chromium-review.googlesource.com/1003995
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Philipp Weiß <phweiss@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>