rockchip: pass thumbnail buffer size correctly in ImgEncoder
With the change of CL:959841, ImgEncoder allocates a larger buffer for
thumbnail compression, but it fails to pass on the buffer size
correctly while converting the buffer type for ImgEncoderCore, which
finds the buffer too small and re-allocates it. Not acknowledging the
re-allocation, ImgEncoder makes EXIF with the wrong thumbnail buffer.

BUG=b:72365664
TEST='test_that scarlet cheets_CTS_N.arm.CtsCameraTestCases'
     and not failed at StillCaptureTest#testJpegExif:
     expected:<320> but was:<-1>'

Change-Id: I59467964453d571036abe8c5631402c75e47446c
Reviewed-on: https://chromium-review.googlesource.com/1004075
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: yichong zhong <zyc@rock-chips.com>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>
Reviewed-by: yichong zhong <zyc@rock-chips.com>