shill: Allow shill user to own flimflam DBUS object.
This CL is in preparation to run the shill process tree as a non-root
user.

Shill will run as user 'shill' and needs to own the flimflam DBUS
object. All code that will run as the shill user currently runs as root,
so this isn't a security regression.

BUG=chromium:649417
TEST=tested as part of larger shill sandboxing debug CL

Change-Id: I2c062f2492e899991300e9f00862018118fcf80d
Reviewed-on: https://chromium-review.googlesource.com/1007483
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>