login: Increase the timeout in AndroidOciWrapper::StartContainer()
The total timeout in session_manager for starting the container
used be 40s (30s for the ::SYNC Upstart signal plus 10s in
AndroidOciWrapper::StartContainer()).

We recently removed the ::SYNC call for running arc-setup, and
instead run arc-setup in run_oci. Because of that, the 10s
timeout needs to be longer.

Although the original total timeout was 40s, it is probably too
long. This CL uses 30s as the new timeout.

BUG=b:77594722
TEST=emerge

Change-Id: I09b1a9eb311984d8a9fdd570149c1773854bf66a
Reviewed-on: https://chromium-review.googlesource.com/1011396
Commit-Ready: Yusuke Sato <yusukes@chromium.org>
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>