chromeos-config: YAML config README.md
Rewrote the config readme to explain the YAML config schema and how to
use it.

Also fixed a minor formatting bug in the generated schema definitions
(when \n are present) and a couple schema issues affecting the generated
readme.

BUG=b:66950157
TEST=unit tests
CQ-DEPEND=CL:1011373

manual updates to the readme file with all of the yaml stuff

Change-Id: I155447376925f87962ae919c6a71a94a0acf5d14
Reviewed-on: https://chromium-review.googlesource.com/1011984
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>