libbrillo: Set default values for RollbackAllowedMilestones policy.
RollbackAllowedMilestones policy values must be between 0 and 4.
Default for enterprise devices is 4, for consumer devices the
value is always 0.

Because DevicePolicy is not available for consumer devices, this
CL adds IsConsumerDevice to PolicyProvider, so update_engine can
check whether it should use the consumer 0 or not set anything if
the policy is not yet loaded.

Not setting anything in the case when the device is still in
OOBE is important because setting a smaller value can permanently
disable rollback to an older version.

BUG=chromium:813036
TEST='cros_run_unit_tests --board=cyan --packages libbrillo'

Change-Id: I228b12ddf3ba05f3230d111ac05d74c0330999c4
Reviewed-on: https://chromium-review.googlesource.com/1013461
Commit-Ready: Marton Hunyady <hunyadym@chromium.org>
Tested-by: Marton Hunyady <hunyadym@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>