login: Don't reboot due to failed UI starting when cros_debug
Since the dawn of time Chrome OS kernel developers have been annoyed
by the fact that the system seems to "randomly" reboot sometimes when
they happen to put on a kernel where UI doesn't come up.

During bringup or during upstreaming, it's pretty common to end up in
a state where the UI doesn't come up.  I personally have experienced
this lots of times.  My typical fix is to 'gzip
/etc/init/ui-respawn.conf', but it's annoying to have to do this all
the time.  Also: many people aren't aware of ui-respawn rebooting the
kernel (and they don't have /var/log/messages open), so to them it
looks like the kernel randomly crashes sometimes (in reality it
"crashes" when it's been long enough since the last reboot).

It seems like a sane fix for all this is to just disable the
auto-rebooting when in debug mode.  This could be a debug build or
could be someone in "dev" mode.  Presumably said people are competent
enough to reboot the machine themselves if the UI isn't coming up.

BUG=None
TEST=Make UI not come up, check /var/log/messages

Change-Id: Ie5af6996267d9fff11e206912b5b2e6dd99518dd
Reviewed-on: https://chromium-review.googlesource.com/1016001
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>