authpolicy: Move preg_parser from libchrome into authpolicy
Moves the .cc, .h, unit test + testdata and fuzzer for preg_parser into
authpolicy/policy, out from libchrome. This is because preg_parser is no
longer used by chrome or libchrome, authpolicy is the only remaining
user.

(This is the authpolicy half of a change that touches libchrome and
authpolicy.)

BUG=chromium:832044
TEST=cros_run_unit_tests --board=amd64-generic --packages authpolicy
CQ-DEPEND=CL:1016763

Change-Id: I8c262384ef4b79de778147c42ce79dd734c77800
Reviewed-on: https://chromium-review.googlesource.com/1016605
Commit-Ready: A Olsen <olsen@chromium.org>
Tested-by: A Olsen <olsen@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>