vm_tools: generalize the 'has output' workaround.
The workaround that exists to ensure that cursor has the correct
scale is also useful for other surface. This moves the logic to
surface commit and applies it to all surfaces that have a role.
E.g. is a cursor or shell surface.

The result is that Chrome OS resolution changes no longer cause
GTK3 applications to drop into low DPI mode.

We can eventually remove this workaround when Exo has been improved
to track outputs for each surface. That's a large task so having
this relatively simple workaround for now is useful.

BUG=none
TEST=gedit is not dropping to low dpi when cros changes resolution

Change-Id: Ic23c4913cfc27423343b83ba479d0829750a8da8
Reviewed-on: https://chromium-review.googlesource.com/1017961
Commit-Ready: David Reveman <reveman@chromium.org>
Tested-by: David Reveman <reveman@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>