cryptohome: Make MountEx two-phased - gen credentials first
Split the MountEx worker job (DoMountEx) into two functions. The first
function only does some input parameters validation and builds the
credentials object, while all state-depending and state-changing mount
operations are moved into the second function.

Both parts are still executed synchronously currently - in order to
preserve the previous semantics of MountEx, which was to work on the
worker thread as a *single* job, which implies that the order of
operations was preserved.

This is preparation for challenge-response based credentials
implementation, for which the credentials object construction will be an
asynchronous operation (involving non-blocking inter-process requests).
The split made in this CL allows to avoid side effects of state changing
during interleaving MountEx and other operations.

BUG=chromium:806788
TEST=existing tests

Change-Id: I311720b38c0e75a0c44a33abf49450db618ca1ac
Reviewed-on: https://chromium-review.googlesource.com/1019406
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>